feat: split up Account2FA and AccountMultisig
