+++ packages/accounts/src/constants.ts
@@ -1,0 +1,10 @@
+import { parseNearAmount } from '@near-js/client-core';
+import BN from 'bn.js';
+
+export const MULTISIG_STORAGE_KEY = '__multisigRequest';
+export const MULTISIG_ALLOWANCE = new BN(parseNearAmount('1'));
+// TODO: Different gas value for different requests (can reduce gas usage dramatically)
+export const MULTISIG_GAS = new BN('100000000000000');
+export const MULTISIG_DEPOSIT = new BN('0');
+export const MULTISIG_CHANGE_METHODS = ['add_request', 'add_request_and_confirm', 'delete_request', 'confirm'];
+export const MULTISIG_CONFIRM_METHODS = ['confirm'];
