+++ packages/accounts/src/index.ts
@@ -7,16 +7,18 @@ export {
     ChangeFunctionCallOptions,
     ViewFunctionCallOptions,
 } from './account';
+export { Account2FA } from './account_2fa';
+export { AccountMultisig } from './account_multisig';
+export { Connection } from './connection';
 export {
-    Account2FA,
-    AccountMultisig,
     MULTISIG_STORAGE_KEY,
     MULTISIG_ALLOWANCE,
     MULTISIG_GAS,
     MULTISIG_DEPOSIT,
     MULTISIG_CHANGE_METHODS,
     MULTISIG_CONFIRM_METHODS,
+} from './constants';
+export {
     MultisigDeleteRequestRejectionError,
     MultisigStateStatus,
-} from './account_multisig';
-export { Connection } from './connection';
+} from './types';
